+++ packages/fighting-design/footer/index.ts
@@ -4,7 +4,9 @@ import { install } from '../_utils'
 
 export const FFooter = install(Footer)
 
-export type FFooterInstance = InstanceType<typeof Footer>
+export type FooterInstance = InstanceType<typeof Footer>
+
+export type { FooterPropsType } from './src/footer'
 
 declare module 'vue' {
   export interface GlobalComponents {